Inject plain text logger to View assembly
