+++ code/triangle/View/build.gradle
@@ -2,6 +2,7 @@ apply from:   rootDir.getAbsolutePath() + '/code/config/scripts/javafx.plugin'
 
 dependencies {
     compile project(':triangle-ViewModel')
+    compile project(':triangle-Infrastructure')
 }
 
 sourceSets {
